feat: jbang enable all the demos and readme
