+++ demos/picocli-demo/src/main/java/dev/tamboui/demo/PicoCLIDemo.java
@@ -1,3 +1,7 @@
+///usr/bin/env jbang "$0" "$@" ; exit $?
+//DEPS dev.tamboui:tamboui-picocli:LATEST
+//DEPS dev.tamboui:tamboui-jline:LATEST
+
 /*
  * Copyright (c) 2025 TamboUI Contributors
  * SPDX-License-Identifier: MIT
